BGMとBPMを設定可能に
