+++ Assets/Audio/Music/tamborine_120bpm_4-4time_301beats_stereo_GrHpg3.mp3.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 574a6e4cff166ff47854af3ad3e21c26
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
