+++ Assets/Resources/Scores.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cccc8d61211e1164a8e8e27750687729
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
