+++ Assets/Resources/Scores/Tutorial.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 35a09483f258cc543bcd327f4ec15328
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
