+++ Assets/Scripts/MusicManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 43c736d44fa87a34f8c9ea10233f59af
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
